removed some extra scripts that i didnt want to release yet.
removed the scripts sccperkochunk.cs as i didn't intend to share that yet. it's part of the game i am working on. also removed 2 water shaders that i had found on the web, all credit to their respective owners for the shaders but the scperkochunk is my version and pretty much a copy of "how to do a minecraft chunk" as per the tutorials here https://www.youtube.com/watch?v=_3ZODVA0yKQ